Fixed some bugs

+++ yatube/posts/views.py
@@ -125,12 +125,14 @@ def follow_index(request):
 @login_required
 def profile_follow(request, username):
     author = get_object_or_404(User, username=username)
-    Follow.objects.create(user=request.user, author=author)
+    if author != request.user:
+        Follow.objects.create(user=request.user, author=author)
     return redirect('posts:follow_index')
 
 
 @login_required
 def profile_unfollow(request, username):
     author = get_object_or_404(User, username=username)
-    Follow.objects.get(user=request.user, author=author).delete()
+    if author != request.user:
+        Follow.objects.get(user=request.user, author=author).delete()
     return redirect('posts:follow_index')
